[Podfile] Add support for initialization from a YAML file

+++ lib/cocoapods-core/podfile.rb
@@ -193,18 +193,42 @@ module Pod
     # @return [Podfile] the generated podfile.
     #
     def self.from_file(path)
-      podfile = Podfile.new(path) do
-        string = File.open(path, 'r:utf-8')  { |f| f.read }
-        # Work around for Rubinius incomplete encoding in 1.9 mode
-        if string.respond_to?(:encoding) && string.encoding.name != "UTF-8"
-          string.encode!('UTF-8')
-        end
+      path = Pathname.new(path)
+      unless path.exist?
+        raise StandardError, "No Podfile exists at path `#{path}`."
+      end
+      string = File.open(path, 'r:utf-8')  { |f| f.read }
+      # Work around for Rubinius incomplete encoding in 1.9 mode
+      if string.respond_to?(:encoding) && string.encoding.name != "UTF-8"
+        string.encode!('UTF-8')
+      end
 
+      case path.extname
+      when ''
+        Podfile.from_ruby(string, path)
+      when '.yaml', '.cocoapods'
+        Podfile.from_yaml(string, path)
+      else
+        raise StandardError, "Unsupported Podfile format `#{path}`."
+      end
+    end
+
+    # Configures a new Podfile from the given ruby string.
+    #
+    # @param  [String] string
+    #         The ruby string which will configure the podfile with the DSL.
+    #
+    # @param  [Pathname] path
+    #         The path from which the Podfile is loaded.
+    #
+    # @return [Podfile] the new Podfile
+    #
+    def self.from_ruby(string, path = nil)
+      podfile = Podfile.new(path) do
         begin
           eval(string, nil, path.to_s)
         rescue Exception => e
-          raise DSLError.new("Invalid `#{path.basename}` file: #{e.message}",
-                             path, e.backtrace)
+          raise DSLError.new("Invalid `#{path.basename}` file: #{e.message}", path, e.backtrace)
         end
       end
       podfile
@@ -212,13 +236,16 @@ module Pod
 
     # Configures a new Podfile from the given hash.
     #
-    # @param  [Hash] the hash which contains the information of the
-    #         Podfile.
+    # @param  [Hash] hash
+    #         The hash which contains the information of the Podfile.
+    #
+    # @param  [Pathname] path
+    #         The path from which the Podfile is loaded.
     #
     # @return [Podfile] the new Podfile
     #
-    def self.from_hash(hash)
-      podfile = Podfile.new
+    def self.from_hash(hash, path = nil)
+      podfile = Podfile.new(path)
       internal_hash = hash.dup
       target_definitions_hash = internal_hash.delete('target_definitions')
       podfile.send(:internal_hash=, internal_hash)
@@ -231,13 +258,16 @@ module Pod
 
     # Configures a new Podfile from the given YAML representation.
     #
-    # @param  [String] the YAML encoded hash which contains the information of
-    #         the Podfile.
+    # @param  [String] yaml
+    #         The YAML encoded hash which contains the information of the
+  #         Podfile.
     #
+    # @param  [Pathname] path
+    #         The path from which the Podfile is loaded.
     #
     # @return [Podfile] the new Podfile
     #
-    def self.from_yaml(yaml)
+    def self.from_yaml(yaml, path = nil)
       hash = YAML.load(yaml)
       from_hash(hash)
     end
